feat: add high score and amount of games played
(Completed step 6.)

In game.html:
* Added divs to display the amount of games played and the high score.

In game.js:
* Added endGame method that will make a POST request to the server to
give the points obtained for the current game.  This method will update
the high score that is displayed on the webpage.
  - The line "this.isEnabled = false;" is moved into the endGame
  method, since it is more appropriate to include it with other code
  that occurs at the end of a game.

In app.py:
* For route "/game", added extra parameters for amount of games played
and the high score to the rendering of game.html, in order to display
those on page load.

* Added a new route "/game/highscore" for POST requests of the game
score. This will update the high score and the amount of games played.
This will return the current high score.